requirements.txt
python-dotenv==1.0.0
black==23.7.0
patool==1.12
loguru==0.7.0